ai-ecommerce/requirements.txt
pandas
numpy
sentence-transformers
supabase
google-generativeai
python-dotenv
fastapi
uvicorn[standard]